pyproject.toml
[tool.poetry]
name = "slack-ai-agent-starter"
version = "0.1.0"
description = "Slack AI Agent Starter Kit"
authors = ["kyaukyuai <y.kakui@gmail.com>"]
repository = "https://github.com/kyaukyuai/slack-ai-agent-starter"
keywords = ["slack", "ai", "bot", "langchain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3.11",
]
packages = [
    { include = "slack_ai_agent", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
slack-bolt = "^1.22.0"
python-dotenv = "^1.0.1"
langchain_anthropic = ">=0.3.6,<0.4.0"
langchain_core = ">=0.3.0,<0.4.0"
langchain_openai = ">=0.3.3,<0.4.0"
langchain_community = ">=0.3.0,<0.4.0"
langgraph-sdk = "^0.1.48"
tavily-python = "^0.5.0"
hupper = "^1.12.1"
langgraph = "^0.2.68"
tiktoken = "^0.9.0"
beautifulsoup4 = "^4.13.1"
slack-sdk = "^3.34.0"
youtube-search = "^2.1.2"
langchain-experimental = "^0.3.4"
firecrawl-py = "^1.12.0"
langchain-arcade = "^1.0.0"
pytz = "^2025.1"
python-pptx = "^1.0.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.2"
pre-commit = "^4.1.0"
mypy = "^1.14.1"
langgraph-cli = {extras = ["inmem"], version = "^0.1.65"}
pytest = "^8.0.0"
pytest-asyncio = "^0.25.0"
pytest-mock = "^3.12.0"
pytest-cov = "^5.0.0"

[tool.mypy]
ignore_missing_imports = true
follow_imports = "silent"
disable_error_code = "import-not-found"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
